refactor: change Packaging
